refactor(ai-employee): split customer-yaml validator into per-section files
CI flagged that the monolithic validator.ts exceeded the 500/75/15
file/function/complexity ceilings from CLAUDE.md's coding standards.
Split it into:

- types.ts (209 lines) — type definitions + accepted-value constants
- helpers.ts (146 lines) — shared isPlainObject, checkRequiredString,
  checkEnum, optionalString/Enum/StringList, secretFindingToError
- sections-identity.ts (247 lines) — schema_version, customer_id,
  practice_areas, machine, users
- sections-personas.ts (404 lines) — personas array with per-entry
  recursion split into checkPersonaSlug/Status/Name/Tone/SendAs/
  ChannelBindings + per-skill checkSkill{Name,Ceiling,Version,Enabled}
- sections-connectors.ts (142 lines) — capability-keyed connectors with
  per-entry checkAdapter/Backend/TokenRef/Scopes
- sections-other.ts (412 lines) — scope, escalation, memory (with
  isolation invariants), and optional voice_library/business_hours/
  logging/pause
- validator.ts (218 lines) — thin orchestrator: raw-scan, structural-
  scan, validateSections, assembleCustomerYaml

No behavior change. All 95 existing tests still pass; the public API
exported from customer-yaml/index.ts is unchanged. Reduces every
function under the 75-line / 15-complexity ceiling and every file
under 500 lines.
